Move store setup to separate function
